Create Digital Delay Timer
The Digital Delay Timer LS7212 is an advanced timing device designed to provide precise control over the delay of electrical signals in various applications. The LS7211/LS7212 are monolithic 
CMOS integrated circuits for generating digitally programmable delays. The delay is controlled by 8 binary weighted inputs, WB0-WB7, in conjunction with an applied clock or oscillator frequency. The programmed time delay manifests itself in the Delay Output (OUT) as a function of the Operating Mode selected by the Mode Select inputs A and B: One-Shot, Delayed Operate, Delayed Release or Dual Delay. The time delay is initiated by a transition of the Trigger Input (TRIG).